refactor: move integration pages to management pages
