+++ apps/nextjs/src/app/[locale]/manage/apps/edit/[id]/page.tsx
