+++ apps/nextjs/src/app/[locale]/manage/apps/new/_app-new-form.tsx
@@ -3,7 +3,6 @@
 import { useCallback } from "react";
 import { useRouter } from "next/navigation";
 
-import type { RouterOutputs } from "@homarr/api";
 import { clientApi } from "@homarr/api/client";
 import {
   showErrorNotification,
@@ -14,24 +13,20 @@ import { useScopedI18n } from "@homarr/translation/client";
 import type { validation, z } from "@homarr/validation";
 
 import { revalidatePathAction } from "~/app/revalidatePathAction";
-import { AppForm } from "../../_form";
+import { AppForm } from "../_form";
 
-interface AppEditFormProps {
-  app: RouterOutputs["app"]["byId"];
-}
-
-export const AppEditForm = ({ app }: AppEditFormProps) => {
-  const t = useScopedI18n("app.page.edit.notification");
+export const AppNewForm = () => {
+  const t = useScopedI18n("app.page.create.notification");
   const router = useRouter();
 
-  const { mutate, isPending } = clientApi.app.update.useMutation({
+  const { mutate, isPending } = clientApi.app.create.useMutation({
     onSuccess: () => {
       showSuccessNotification({
         title: t("success.title"),
         message: t("success.message"),
       });
-      void revalidatePathAction("/apps").then(() => {
-        router.push("/apps");
+      void revalidatePathAction("/manage/apps").then(() => {
+        router.push("/manage/apps");
       });
     },
     onError: () => {
@@ -44,23 +39,19 @@ export const AppEditForm = ({ app }: AppEditFormProps) => {
 
   const handleSubmit = useCallback(
     (values: z.infer<typeof validation.app.manage>) => {
-      mutate({
-        id: app.id,
-        ...values,
-      });
+      mutate(values);
     },
-    [mutate, app.id],
+    [mutate],
   );
 
   const submitButtonTranslation = useCallback(
-    (t: TranslationFunction) => t("common.action.save"),
+    (t: TranslationFunction) => t("common.action.create"),
     [],
   );
 
   return (
     <AppForm
       submitButtonTranslation={submitButtonTranslation}
-      initialValues={app}
       handleSubmit={handleSubmit}
       isPending={isPending}
     />
