+++ apps/nextjs/src/app/[locale]/manage/apps/new/page.tsx
