+++ apps/nextjs/src/app/[locale]/manage/integrations/_integration-avatar.tsx
