+++ apps/nextjs/src/app/[locale]/manage/integrations/_integration-secret-inputs.tsx
