+++ apps/nextjs/src/app/[locale]/manage/integrations/edit/[id]/page.tsx
