+++ apps/nextjs/src/app/[locale]/manage/integrations/new/_integration-new-form.tsx
