+++ apps/nextjs/src/app/[locale]/manage/integrations/new/page.tsx
