+++ apps/nextjs/src/app/[locale]/manage/integrations/page.tsx
